Added basic multi page so home page will pprovide option between stopwatch and timer, then can go between the two with buttons thereafter
