+++ stopwatch/stopwatchApp.swift
@@ -12,6 +12,9 @@ struct stopwatchApp: App {
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .sheet(isPresented: .constant(false)) {
+                    TimerView()
+                }
         }
     }
 }
